backend/requirements.txt
# requirements.txt
fastapi==0.136.0
uvicorn==0.44.0
sqlalchemy==2.0.49
python-jose==3.5.0
bcrypt==4.1.2
passlib==1.7.4
email-validator==2.3.0
python-multipart==0.0.6
argon2-cffi==23.1.0
ollama==0.6.1
python-dotenv==1.2.2
psycopg2-binary==2.9.12

